+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Display list of all previously set Gitlab URL Tokens in settings and during the addition of a new remote to the 
+remotes table.
 
 ## [1.10.0] - 2023-04-27